requirements.txt
# PYTHON 3.13
aiohttp==3.13.2
azure-cosmos==4.14.2
azure-storage-blob==12.27.1
dotenv==0.9.9 
fastapi==0.122.0
gunicorn==23.0.0
pydantic==2.12.4
pydantic-settings==2.12.0
python-dotenv==1.2.1
uvicorn==0.38.0
